Save and watch messages in <folder>/hhh/{cur,new}
This reduces the size of individual directories which have to be
rescanned when the modification time changes.  hhh are hexadecimal
digits derived from the CRC-16 value of the Message-Id (currently
11-bits), but the rest of the algorithm doesn't care it is.
It will still find files residing in <folder>/{cur,new} as well.

As the hhh/{cur,new} folders may not exist initially, we need to
incrementally add them to the inotify watch list once we notice that
they exist.  That also means we need to keep track of the paths that are
being watched, and notice if they become unwatched.