Fix touching downloaded message files to the internal date.
We set incorrect timestamps on downloaded message files because, when we
convert from a date_time to a time_t value, we expected mktime() to take
the `struct tm' time zone field into account.  It does not, so all the
other `struct tm' fields were interpreted in the local time zone.